Aspect Ration Preservation
I've added Aspect Ration Preservation to the Compressor. This allows you to scale the image down to the smallest without while maintaining the aspect ration. You can use this feature by setting the width or height to auto. It is as simple as that. Setting both of them to auto will generate an error.